refactor: move implementations to metadata package
