+++ jnosql-lite/mapping-lite-processor/src/main/java/org/eclipse/jnosql/lite/mapping/metadata/LiteClassScanner.java
@@ -12,7 +12,7 @@
  *
  *   Otavio Santana
  */
-package org.eclipse.jnosql.lite.mapping;
+package org.eclipse.jnosql.lite.mapping.metadata;
 
 import jakarta.data.repository.DataRepository;
 import org.eclipse.jnosql.mapping.metadata.ClassScanner;
