+++ jnosql-lite/mapping-lite-processor/src/main/resources/META-INF/services/org.eclipse.jnosql.mapping.metadata.ClassConverter
@@ -1,1 +1,1 @@
-org.eclipse.jnosql.lite.mapping.LiteClassConverter
+org.eclipse.jnosql.lite.mapping.metadata.LiteClassConverter
