+++ jnosql-lite/mapping-lite-processor/src/main/resources/META-INF/services/org.eclipse.jnosql.mapping.metadata.ClassScanner
@@ -1,1 +1,1 @@
-org.eclipse.jnosql.lite.mapping.LiteClassScanner
+org.eclipse.jnosql.lite.mapping.metadata.LiteClassScanner
