+++ jnosql-lite/mapping-lite-processor/src/main/resources/META-INF/services/org.eclipse.jnosql.mapping.metadata.ConstructorBuilderSupplier
@@ -1,1 +1,1 @@
-org.eclipse.jnosql.lite.mapping.LiteConstructorBuilderSupplier
+org.eclipse.jnosql.lite.mapping.metadata.LiteConstructorBuilderSupplier
