+++ jnosql-lite/mapping-lite-processor/src/test/java/org/eclipse/jnosql/lite/mapping/LiteClassScannerTest.java
@@ -15,6 +15,7 @@
 package org.eclipse.jnosql.lite.mapping;
 
 import jakarta.data.repository.DataRepository;
+import org.eclipse.jnosql.lite.mapping.metadata.LiteClassScanner;
 import org.junit.jupiter.api.Test;
 
 import java.util.Set;
