+++ jnosql-lite/mapping-lite-processor/src/test/java/org/eclipse/jnosql/lite/mapping/LiteConstructorBuilderSupplierTest.java
@@ -14,6 +14,7 @@
  */
 package org.eclipse.jnosql.lite.mapping;
 
+import org.eclipse.jnosql.lite.mapping.metadata.LiteConstructorBuilderSupplier;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
